% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_norm.R
\name{quant_norm}
\alias{quant_norm}
\title{Quantile Normalization}
\usage{
quant_norm(gene_exp_mat)
}
\arguments{
\item{gene_exp_mat}{This gene matrix is assumed to be sum normalized but otherwise a regular sparse matrix}
}
\value{
Returns a sparse matrix of the same dimension
}
\description{
This function performs the quantile normalization on a gene expression matrix. Some warning of this function: May run into memory issue
on personal machine if matrix is too large. If needed subsample original matrix and continue operation.
}
\keyword{gene}
\keyword{selection}
