% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_norm.R
\name{sum_norm}
\alias{sum_norm}
\title{Sum Normalization}
\usage{
sum_norm(sparse_mat)
}
\arguments{
\item{sparse_mat}{Takes in a sparse gene expression matrix}
}
\value{
Returns a sparse matrix
}
\description{
This function normalizes the columns of the gene expression matrix, so that each column sums to one. In other words all
the genes in each samples sums to one
}
\keyword{normalization}
